feat: Configure morgan to show data sent in HTTP  requests
The code changes modify the configuration of the morgan middleware to include the data sent in HTTP POST requests. This is achieved by adding a custom token `post-body` that returns the JSON stringified version of the request body. The updated configuration is applied to the Express app using the `app.use` method.